updates to Bootstrap 3

+++ app/helpers/action_button_helper.rb
@@ -1,10 +1,16 @@
 module ActionButtonHelper
-  def action_buttons_for(object)
+  def action_buttons_for(object, options = {})
+    displayed_buttons = %i(show edit destroy)
+    displayed_buttons &= Array(options[:only]) if options[:only]
+    displayed_buttons -= Array(options[:except]) if options[:except]
+
     content_tag :div, class: 'btn-toolbar' do
-      with_options namespace: :admin, size: :mini do |buttons|
-        buttons.view_button_link_for(object) <<
-        buttons.update_button_link_for(object) <<
-        buttons.destroy_button_link_for(object)
+      with_options options.merge(namespace: :admin, size: :mini) do |buttons|
+        html = ""
+        html << buttons.view_button_link_for(object) if displayed_buttons.include?(:show)
+        html << buttons.update_button_link_for(object) if displayed_buttons.include?(:edit)
+        html << buttons.destroy_button_link_for(object) if displayed_buttons.include?(:destroy)
+        html.html_safe
       end
     end
   end
@@ -12,7 +18,8 @@ module ActionButtonHelper
   def create_button_link_for(model, options = {})
     options = options.dup
     namespace = options.delete(:namespace)
-    url = options.delete(:url) { [:new, namespace, model.model_name.singular.to_sym].flatten.compact }
+    parent = Array(options.delete(:parent))
+    url = options.delete(:url) { [:new, namespace, parent, model.model_name.singular.to_sym].flatten.compact }
     options[:type] = :success
     button_link_to prepend_icon(:plus, translate('helpers.links.new', model: model.model_name.human)), url, options
   end
@@ -20,7 +27,8 @@ module ActionButtonHelper
   def destroy_button_link_for(object, options = {})
     options = options.dup
     namespace = options.delete(:namespace)
-    url = options.delete(:url) { [namespace, object].flatten.compact }
+    parent = Array(options.delete(:parent))
+    url = options.delete(:url) { [namespace, parent, object].flatten.compact }
     options[:type] = :danger
     options[:method] = :delete
     options[:data] ||= {}
@@ -31,14 +39,16 @@ module ActionButtonHelper
   def update_button_link_for(object, options = {})
     options = options.dup
     namespace = options.delete(:namespace)
-    url = options.delete(:url) { [:edit, namespace, object].flatten.compact }
+    parent = Array(options.delete(:parent))
+    url = options.delete(:url) { [:edit, namespace, parent, object].flatten.compact }
     button_link_to prepend_icon(:edit, translate('helpers.links.edit', model: object.class.model_name.human)), url, options
   end
 
   def view_button_link_for(object, options = {})
     options = options.dup
     namespace = options.delete(:namespace)
-    url = options.delete(:url) { [namespace, object].flatten.compact }
+    parent = Array(options.delete(:parent))
+    url = options.delete(:url) { [namespace, parent, object].flatten.compact }
     options[:type] = :primary
     button_link_to prepend_icon(:eye_open, translate('helpers.links.view', model: object.class.model_name.human)), url, options
   end
